Simplify block fast-fail validation flow
Move shared generator-reference checks into a neutral consensus helper and focus regression coverage on the new resource-exhaustion boundaries.
