+++ chia/consensus/block_body_validation.py
@@ -23,41 +23,12 @@ from chiabip158 import PyBIP158
 from chia.consensus.block_rewards import calculate_base_farmer_reward, calculate_pool_reward
 from chia.consensus.blockchain_interface import BlockRecordsProtocol
 from chia.consensus.coinbase import create_farmer_coin, create_pool_coin
+from chia.consensus.generator_validation import validate_generator_ref_list
 from chia.types.blockchain_format.coin import Coin, hash_coin_ids
 from chia.util.errors import Err
 from chia.util.hash import std_hash
 
 log = logging.getLogger(__name__)
-
-
-def validate_generator_ref_list(
-    constants: ConsensusConstants,
-    block: FullBlock | UnfinishedBlock,
-    height: uint32,
-    prev_transaction_block_height: uint32,
-) -> Err | None:
-    assert block.transactions_info is not None
-
-    if block.transactions_generator_ref_list == []:
-        if block.transactions_info.generator_refs_root != bytes([1] * 32):
-            return Err.INVALID_TRANSACTIONS_GENERATOR_REFS_ROOT
-        return None
-
-    if prev_transaction_block_height >= constants.SOFT_FORK9_HEIGHT:
-        return Err.TOO_MANY_GENERATOR_REFS
-    if block.transactions_generator is None:
-        return Err.INVALID_TRANSACTIONS_GENERATOR_REFS_ROOT
-    if len(block.transactions_generator_ref_list) > constants.MAX_GENERATOR_REF_LIST_SIZE:
-        return Err.TOO_MANY_GENERATOR_REFS
-    generator_refs_hash = std_hash(b"".join(i.stream_to_bytes() for i in block.transactions_generator_ref_list))
-    if block.transactions_info.generator_refs_root != generator_refs_hash:
-        return Err.INVALID_TRANSACTIONS_GENERATOR_REFS_ROOT
-    if any(index >= height for index in block.transactions_generator_ref_list):
-        return Err.FUTURE_GENERATOR_REFS
-
-    return None
-
-
 #  peak->  o
 #  main    |
 #  chain   o  o <- peak_height  \ additions and removals
